==> 01_Preproc_DCT.ipynb <==
# Data preprocessing script - DataCollectionForms

Authors: [Andreas Sørensen](https://www.linkedin.com/in/a-soerensen) and [Martin Röck](https://www.linkedin.com/in/martinroeck/) - Source: https://doi.org/10.5281/zenodo.5895051

# SETUP
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import time

# set data frame printing settings
pd.set_option("display.max_rows", None, "display.max_columns", None)

print("Pre-processing DCT data...")

# DCT all input data

# IMPORT

#File name/path
filename = '00_data/0_data_input_raw/EU-ECB_DataCollectionForm_all_data_input.xlsx'

#Load individual sheets
df_sheet = pd.read_excel(filename, sheet_name='DataCollectionTable')

#Recreate dataframe with parameter code as header instead of first header (delete first three rows)
#This will also assign datatypes correctly
headers = df_sheet.iloc[2]
df_DCT = pd.DataFrame(df_sheet.values[3:], columns=headers)

#Drop first column
df_DCT = df_DCT.iloc[:,1:]

#Drop all rows which have only missing values (NaN)
df_DCT.dropna(how='all',inplace=True)
#Fill NaN entries with n/a for completeness
df_DCT.fillna('n/a',inplace=True)

df_DCT_processed = df_DCT

#Create CSV
filename = '00_data/1_data_pre_processed/DCT_processed.csv'
df_DCT_processed.to_csv(filename, index=False)

print("DCT data pre-processed.")
